fix: 修复 HITL 创建 SQL 参数类型推断失败
INSERT...SELECT 中未转型的 $2 在 node-pg 下会与 ::varchar 空值检查冲突；为 HITL/execution 相关语句补显式转型，并为 DB 错误附带 sql 摘要。

+++ src/adapter/postgres/PostgresSessionRepository.ts
@@ -85,7 +85,7 @@ export class PostgresSessionRepository implements SessionRepository {
     const safeLimit = Math.max(1, Math.min(100, Math.trunc(limit)));
     const safeOffset = Math.max(0, Math.trunc(offset));
     const result = await this.db.query(
-      `SELECT * FROM sessions WHERE user_id = $1 ORDER BY updated_at DESC LIMIT $2 OFFSET $3`,
+      `SELECT * FROM sessions WHERE user_id = $1 ORDER BY updated_at DESC LIMIT $2::int OFFSET $3::int`,
       { 1: this.userId, 2: safeLimit, 3: safeOffset }
     );
     return result.rows.map((r) => this.rowToMeta(r));
